+++ mdmp/group_analysis/is/__init__.py
@@ -13,8 +13,8 @@ or::
 
 from .aggregation import (
     GlobalBetaMCResult,
-    ISAggregateOptions,
     ISAggregatedMDMView,
+    ISAggregateOptions,
     ISAggregationResult,
     MCContributorMode,
     MCPosteriorSource,